refactor(cli): extract get schema
